requirements.txt

Django==3.2.8
daphne==3.0.2
channels==3.0.4
whitenoise==5.3.0
python-engineio==4.2.1
python-socketio==5.4.1
numpy==1.21.5
scipy==1.7.3
matplotlib==3.5.1
